Reconfigured post pin UI flow
Now driven by buttons as per the spec instead of moving forward on return from text fields.

Got rid of LocationTextFieldDelegate, brough logic into AddLocationVC class — logic was too closely tied to the VC to be compartmentalized in a separate class.

Moved butons to top of screen on location and subtitle screens, negating the need to adjust the screen for the keyboard.

Added alert error notices to user when the location is not properly geocoded, and when the link is empty or invalid. The user can now opt to post an invalid or empty link after being notified.